==> 02.gen_question_feat.ipynb <==
import pandas as pd
import numpy as np
import pickle
import gc
import os
import time
import copy
import multiprocessing as mp
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from sklearn.feature_extraction.text import CountVectorizer
from tqdm import tqdm
from scipy import sparse, spatial
import warnings
warnings.filterwarnings('ignore')
tqdm.pandas()

tic = time.time()
DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'
if not os.path.exists(SAVE_PATH):
    print('create dir: %s' % SAVE_PATH)
    os.mkdir(SAVE_PATH)

print('gen_question_feat...')

data = pd.read_pickle(os.path.join(DATA_PATH, 'invite_data.pkl'))
question_info = pd.read_pickle(os.path.join(DATA_PATH, 'question_info.pkl'))

# title、desc词计数，topic计数
question_info['num_title_sw'] = question_info['title_sw'].progress_apply(len)
question_info['num_title_w'] = question_info['title_w'].progress_apply(len)
question_info['num_desc_sw'] = question_info['desc_sw'].progress_apply(len)
question_info['num_desc_w'] = question_info['desc_w'].progress_apply(len)
question_info['num_qtopic'] = question_info['topic'].progress_apply(len)

feats = ['qid', 'qtime', 'num_title_sw', 'num_title_w', 'num_desc_sw', 'num_desc_w', 'num_qtopic', 'qhour']
question_feat = question_info[feats]

data = data.merge(question_feat, 'left', 'qid')
data['inv_que_gap'] = data['itime'] - data['qtime']
feats += ['inv_que_gap']
question_feat = data[feats].drop(columns=['qid', 'qtime'])
question_feat.head(2)

question_feat.to_pickle(os.path.join(SAVE_PATH, 'question_feat.pkl'))

toc = time.time()
print('Used time: %d' % int(toc-tic))
