==> 07.gen_prev_ans_times_stat.ipynb <==
import pandas as pd
import numpy as np
import pickle
import gc
import os
import time
import copy
import multiprocessing as mp
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from sklearn.feature_extraction.text import CountVectorizer
from tqdm import tqdm
from scipy import sparse, spatial
import warnings
warnings.filterwarnings('ignore')
tqdm.pandas()

tic = time.time()
DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'
if not os.path.exists(SAVE_PATH):
    print('create dir: %s' % SAVE_PATH)
    os.mkdir(SAVE_PATH)

print('gen_prev_ans_times_stat...')

inv_last_answer_time = pd.read_pickle(os.path.join(DATA_PATH, 'inv_last_answer_time.pkl'))
answer_info = pd.read_pickle(os.path.join(DATA_PATH, 'answer_info_tiny.pkl'))

answer_info['user_time'] = answer_info['uid'].astype(str) + '_' + answer_info['atime'].astype(str)
answer_info.sort_values(by=['uid', 'user_time'],inplace=True)
answer_info = answer_info[['uid', 'user_time', 'atime']].reset_index(drop=True)

answer_info.head()

prev_ans_times = []
times = []
last = None
for a, _, t in tqdm(answer_info.values):
    if last is None or last != a:
        times = [t]
    else:
        times.append(t)
    prev_ans_times.append(list(times))
    last = a

answer_info['prev_ans_times'] = prev_ans_times
answer_info.drop_duplicates(subset='user_time', keep='last', inplace=True)

min_ = []
mean_ = []
std_ = []
for var in tqdm(answer_info['prev_ans_times']):
    min_.append(np.min(var))
    mean_.append(np.mean(var))
    std_.append(np.std(var))  

tmp = answer_info[['user_time']]
tmp['prev_ans_times_min'] = min_
tmp['prev_ans_times_mean'] = mean_
tmp['prev_ans_times_std'] = std_

tmp = inv_last_answer_time.merge(tmp, 'left', 'user_time')
tmp.head(2)

tmp[['prev_ans_times_' + st for st in ['min', 'mean', 'std']]].to_pickle(os.path.join(SAVE_PATH, 'prev_ans_times_st.pkl'))

toc = time.time()
print('Used time: %d' % int(toc-tic))
