==> 11.gen_topic_sim_stats.ipynb <==
import pandas as pd
import numpy as np
import pickle
import gc
import os
import time
import copy
import multiprocessing as mp
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from sklearn.feature_extraction.text import CountVectorizer
from tqdm import tqdm
from scipy import sparse, spatial
import warnings
warnings.filterwarnings('ignore')
tqdm.pandas()

tic = time.time()
DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'
if not os.path.exists(SAVE_PATH):
    print('create dir: %s' % SAVE_PATH)
    os.mkdir(SAVE_PATH)

print('gen_topic_sim_stats...')

data = pd.read_pickle(os.path.join(DATA_PATH, 'invite_data.pkl'))
user_info = pd.read_pickle(os.path.join(DATA_PATH, 'user_info.pkl'))
question_info = pd.read_pickle(os.path.join(DATA_PATH, 'question_info.pkl'))

invite_id = data[['uid', 'qid']]
invite_id_qm = invite_id.merge(user_info[['uid', 'topic_a', 'topic_ik']], 'left', 'uid').merge(question_info[['qid', 'topic']], 'left', 'qid')

invite_id_qm.head()

topic_comb_sim_mat = np.load('../pkl/topic_comb_sim_mat.npy')

def get_topic_sim(i, j):
    if i == 0 or j == 0:
        return 0
    if i == j:
        return 1
    if i > j:
        return topic_comb_sim_mat[j-1][i-1]
    else:
        return topic_comb_sim_mat[i-1][j-1]

## 用户问题topic相似度统计

def process(df):
    topic_qu_sims_st = np.zeros((len(df), 8))
    for i, (tas, tis, ts) in enumerate(tqdm(df[['topic_a', 'topic_ik', 'topic']].values)):
        ass = []
        iss = []
        for t in ts:
            for ta in tas:
                ass.append(get_topic_sim(t, ta))
            for ti in tis:
                iss.append(get_topic_sim(t, ti))
        ass = ([0] if len(ass) == 0 else ass)
        iss = ([0] if len(iss) == 0 else iss)
        topic_qu_sims_st[i] = np.array([np.min(ass), np.max(ass), np.mean(ass), np.std(ass),
                                        np.min(iss), np.max(iss), np.mean(iss), np.std(iss)])
    return topic_qu_sims_st

with mp.Pool(8) as pool:
    ret = pool.map(process, np.array_split(invite_id_qm, 8))
ret = np.vstack(ret)

cols = ['topic_%s_sims_%s' % (t, st) for t in ['a', 'i'] for st in ['min', 'max', 'mean', 'std']]
topic_qu_sims_stat = pd.DataFrame(ret, columns=cols)
print(topic_qu_sims_stat.shape)

topic_qu_sims_stat.to_pickle(os.path.join(SAVE_PATH, 'topic_qu_sims_stat.pkl'))
topic_qu_sims_stat.head(2)

## 用户历史回答问题和当前问题 topic 相似度统计

prev_ans_ques = pd.read_pickle(os.path.join(DATA_PATH, 'prev_ans_ques.pkl'))[['prev_ans_ques']]

question_info = question_info.set_index('qid')
question_info.head(2)

invite_id = pd.concat([invite_id, prev_ans_ques], axis=1)
invite_id.head(3)

def process(df):
    prev_topic_sims_st = np.zeros((len(df), 4))
    for i, (q1, q2s) in enumerate(tqdm(df[['qid', 'prev_ans_ques']].values)):
        t1s = question_info.loc[q1, 'topic']
        t2s = []
        if type(q2s) == list:
            for q2 in q2s:
                ts = question_info.loc[q2, 'topic']
                if type(ts) == list:
                    t2s += ts
        ss = []
        for t1 in t1s:
            for t2 in t2s:
                ss.append(get_topic_sim(t1, t2)) 
        ss = ([0] if len(ss) == 0 else ss)
        prev_topic_sims_st[i] = np.array([np.min(ss), np.max(ss), np.mean(ss), np.std(ss)])
    return prev_topic_sims_st

with mp.Pool(8) as pool:
    ret = pool.map(process, np.array_split(invite_id, 8))
ret = np.vstack(ret)

cols = ['prev_topic_sims_%s' % (st) for st in ['min', 'max', 'mean', 'std']]
prev_topic_sims_stat = pd.DataFrame(ret, columns=cols)
print(prev_topic_sims_stat.shape)

prev_topic_sims_stat.to_pickle(os.path.join(SAVE_PATH, 'prev_topic_sims_stat.pkl'))
prev_topic_sims_stat.head(5)

toc = time.time()
print('Used time: %d' % int(toc-tic))
