==> 32.user_ques_topic_dnn_enc_A.ipynb <==
import pandas as pd
import numpy as np
import pickle
import gc
import os
import time
import copy
import multiprocessing as mp
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import roc_auc_score, log_loss
from tqdm import tqdm
from scipy import sparse, spatial
import warnings
warnings.filterwarnings('ignore')
tqdm.pandas()

import os
import tensorflow as tf

os.environ['CUDA_VISIBLE_DEVICES'] = "0"
config = tf.compat.v1.ConfigProto()
# config = tf.ConfigProto()
config.gpu_options.allow_growth = True
session = tf.compat.v1.Session(config=config)
# session = tf.Session(config=config)

from keras.models import Sequential, Model
from keras.layers import Input, Dense, CuDNNLSTM, Bidirectional, Embedding, CuDNNGRU, Conv1D, MaxPooling1D, AveragePooling1D
from keras.layers import Flatten, PReLU, Dropout, BatchNormalization, SpatialDropout1D, concatenate
from keras import initializers, regularizers, constraints, optimizers, layers
from keras.preprocessing.sequence import pad_sequences
from keras.utils import Sequence
from keras.callbacks import Callback
from keras import backend as K
from keras.engine.topology import Layer 
# from keras.utils.training_utils import multi_gpu_model

tic = time.time()
DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'
WEIGHT_PATH = './weights/'
if not os.path.exists(SAVE_PATH):
    print('create dir: %s' % SAVE_PATH)
    os.mkdir(SAVE_PATH)
if not os.path.exists(WEIGHT_PATH):
    print('create dir: %s' % WEIGHT_PATH)
    os.mkdir(WEIGHT_PATH)

print('user_ques_topic_dnn_enc_A...')

data = pd.read_pickle(os.path.join(DATA_PATH, 'invite_data.pkl'))
user_info = pd.read_pickle(os.path.join(DATA_PATH, 'user_info.pkl'))
question_info = pd.read_pickle(os.path.join(DATA_PATH, 'question_info.pkl'))
topic = pd.read_pickle(os.path.join(DATA_PATH, 'topic.pkl'))
len_train = (~data['label'].isnull()).sum()
print(len_train)

data.head()

user_info['utopic'] = user_info['topic_a'] + user_info['topic_ik']
question_info.rename(columns={'topic': 'qtopic'}, inplace=True)

data = data.merge(user_info[['uid', 'utopic']], 'left', 'uid').merge(question_info[['qid', 'qtopic']], 'left', 'qid')
# data['topic'] = data['topic'].fillna('-1')
data.head()

max_len_u = data['utopic'].apply(len).max()
max_len_q = data['qtopic'].apply(len).max()
print(max_len_u, max_len_q) # 110

np.random.seed(42)
def shuffle(l):
    np.random.shuffle(l)
    return l

data['utopic'] = data['utopic'].progress_apply(shuffle)
data['qtopic'] = data['qtopic'].progress_apply(shuffle)

data.head()

qid_topic_map_32 = data[['qid', 'qtopic']].copy()
qid_topic_map_32.drop_duplicates(subset=['qid'], inplace=True)
qid_topic_map_32.to_pickle('./shuffle_map/qid_topic_map_32.pkl')
print(qid_topic_map_32.shape)

uid_topic_map_32 = data[['uid', 'utopic']].copy()
uid_topic_map_32.drop_duplicates(subset=['uid'], inplace=True)
uid_topic_map_32.to_pickle('./shuffle_map/uid_topic_map_32.pkl')
print(uid_topic_map_32.shape)

utopics = pad_sequences(data['utopic'], maxlen=max_len_u)
qtopics = pad_sequences(data['qtopic'], maxlen=max_len_q)

utopics.shape

embed_weights = np.array([[0] * 64] + [list(v) for v in topic['vector'].values])

embed_weights.shape

class DataSequence(Sequence):
    
    def __init__(self, xu, xq, y, batch_size=128):
        self.xu = xu
        self.xq = xq
        self.y = y
        self.batch_size = batch_size
         
    def __len__(self):
        return int(np.ceil(self.xu.shape[0] / float(self.batch_size)))
    
    def __getitem__(self, idx):
        batch_idx = np.arange(idx * self.batch_size, min((idx + 1) * self.batch_size, self.xu.shape[0]))
        batch_xu = self.xu[batch_idx]
        batch_xq = self.xq[batch_idx]
        batch_y = self.y[batch_idx]
        return [batch_xu, batch_xq], batch_y

class MetricsCallback(Callback):
    def __init__(self, x_trn_u, x_trn_q, y_trn, x_val_u, x_val_q, y_val, batch_size=128, save_name='weight.h5'):
        self.trn_generator = DataSequence(x_trn_u, x_trn_q, y_trn, BATCH_SIZE)
        self.val_generator = DataSequence(x_val_u, x_val_q, y_val, BATCH_SIZE)
        self.y_trn = y_trn
        self.y_val = y_val
        self.save_name = save_name
        self.best_score = 0.5

    def on_epoch_end(self, epoch, logs={}):
        # eval train
        y_pred = self.model.predict_generator(self.trn_generator, 
                                              max_queue_size=10, 
                                              workers=1, 
                                              use_multiprocessing=False, 
                                              verbose=0)
        roc = roc_auc_score(self.y_trn, y_pred)
        # eval valid
        y_pred_val = self.model.predict_generator(self.val_generator, 
                                              max_queue_size=10, 
                                              workers=1, 
                                              use_multiprocessing=False, 
                                              verbose=0)
        roc_val = roc_auc_score(self.y_val, y_pred_val)
        print('\rroc-auc: %s - roc-auc_val: %s' % (str(round(roc,4)),str(round(roc_val,4))),end=100*' '+'\n')
#         print(roc_val)
#         print(log_loss(self.y_val, y_pred_val)/len(y_pred_val))
        if roc_val > self.best_score:
            self.best_score = roc_val
            self.model.save_weights(os.path.join(WEIGHT_PATH, self.save_name))
        
        return

def Net(hiddens=(256,)):
    # user topic
    inp_u = Input(shape=(max_len_u,))
    x = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_u)
    x = Dropout(0.2)(x)
    x = Conv1D(256, 2, padding='valid', activation='relu', strides=1)(x)
    x = AveragePooling1D(pool_size=4)(x)
    x = Flatten()(x)
    # ques topic
    inp_q = Input(shape=(max_len_q,))
    y = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_q)
    y = Dropout(0.2)(y)
    y = Conv1D(256, 2, padding='valid', activation='relu', strides=1)(y)
    y = AveragePooling1D(pool_size=4)(y)
    y = Flatten()(y)
    # concat
    x = concatenate([x, y], axis=1)
    for hs in hiddens:
        x = Dense(hs)(x)
        x = BatchNormalization()(x)
        x = PReLU()(x)
        x = Dropout(0.2)(x)
    out = Dense(1, activation="sigmoid")(x)
    model = Model(inputs=[inp_u, inp_q], outputs=out)
    return model

train_x_u = utopics[:len_train]
train_x_q = qtopics[:len_train]
test_x_u = utopics[len_train:]
test_x_q = qtopics[len_train:]
train_y = data['label'].values[:len_train]

print(train_x_u.shape)
print(test_x_u.shape)
print(train_y.sum() / len(train_y))

BATCH_SIZE = 1024

kfold = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

# for i, (tr_idx, va_idx) in enumerate(kfold.split(train_x_u, train_y)):
#     print('-'*100)
#     print('Fold %d' % i)
#     trn_x_u, trn_x_q, trn_y = train_x_u[tr_idx], train_x_q[tr_idx], train_y[tr_idx]
#     val_x_u, val_x_q, val_y = train_x_u[va_idx], train_x_q[va_idx], train_y[va_idx]
    
#     K.clear_session()
#     model = Net(hiddens=(256,256,256))
#     model.compile(loss='binary_crossentropy', optimizer='adam',)
    
#     trn_generator = DataSequence(trn_x_u, trn_x_q, trn_y, batch_size=BATCH_SIZE)
#     val_generator = DataSequence(val_x_u, val_x_q, val_y, batch_size=BATCH_SIZE)
    
#     history = model.fit_generator(generator=trn_generator, 
#                         epochs=30, 
#                         verbose=1, 
#                         callbacks=[MetricsCallback(trn_x_u, trn_x_q, trn_y, val_x_u, val_x_q, val_y, 
#                                                    batch_size=BATCH_SIZE*4, 
#                                                    save_name='user_ques_topic_dnn_enc_A_weight_fold%d.h5' % i)], 
#                         max_queue_size=100, 
#                         workers=1, 
#                         use_multiprocessing=False,)
    
#     del trn_x_u, trn_x_q, trn_y, val_x_u, val_x_q, val_y
#     gc.collect()

# 0.4608

test_generator = DataSequence(test_x_u, test_x_q, np.zeros(test_x_u.shape[0]), batch_size=BATCH_SIZE * 4)

test_pred = np.zeros((test_x_u.shape[0], 1))
train_pred = np.zeros((train_x_u.shape[0], 1))

for i, (tr_idx, va_idx) in enumerate(kfold.split(train_x_u, train_y)):
    print('-'*100)
    print('Fold %d' % i)
    trn_x_u, trn_x_q, trn_y = train_x_u[tr_idx], train_x_q[tr_idx], train_y[tr_idx]
    val_x_u, val_x_q, val_y = train_x_u[va_idx], train_x_q[va_idx], train_y[va_idx]
    
    K.clear_session()
    model = Net(hiddens=(256,256,256))
    model.compile(loss='binary_crossentropy', optimizer='adam',)
    
    trn_generator = DataSequence(trn_x_u, trn_x_q, trn_y, batch_size=BATCH_SIZE)
    val_generator = DataSequence(val_x_u, val_x_q, val_y, batch_size=BATCH_SIZE)
    
    model.load_weights(os.path.join(WEIGHT_PATH, 'user_ques_topic_dnn_enc_A_weight_fold%d.h5' % i))
    
    train_pred[va_idx] = model.predict_generator(val_generator)
    test_pred += model.predict_generator(test_generator) / 5
    
    print(roc_auc_score(val_y, train_pred[va_idx]))
    
    del trn_x_u, trn_x_q, trn_y, val_x_u, val_x_q, val_y
    gc.collect()
    print('Predict Done.')

arr = np.vstack([train_pred, test_pred])
feat = pd.DataFrame(arr, columns=['user_ques_topic_dnn_enc_A'])
feat.head()

feat.to_pickle('./feats/user_ques_topic_dnn_enc_A.pkl')

feat.shape
