==> 34.ques_user_title_desc_textattbirnn_cossim_enc.ipynb <==
import pandas as pd
import numpy as np
import pickle
import gc
import os
import time
import copy
import multiprocessing as mp
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics import roc_auc_score, log_loss
from tqdm import tqdm
from scipy import sparse, spatial
import warnings
warnings.filterwarnings('ignore')
tqdm.pandas()

# !pip install catboost==0.19.1
# !pip install pandas==0.25.3
# !pip install scikit-learn==0.19.2
# !pip install numpy==1.17.4 
# !pip install tensorflow-gpu==1.13.1
# !pip install keras==2.3.1
# !pip install -U tqdm

import os
import tensorflow as tf

os.environ['CUDA_VISIBLE_DEVICES'] = "0"
# config = tf.compat.v1.ConfigProto()
config = tf.ConfigProto()
config.gpu_options.allow_growth = True
# session = tf.compat.v1.Session(config=config)
session = tf.Session(config=config)

from keras.models import Sequential, Model
from keras.layers import Input, Dense, CuDNNLSTM, Bidirectional, Embedding, CuDNNGRU, Conv1D, MaxPooling1D, AveragePooling1D, GlobalMaxPooling1D
from keras.layers import Flatten, PReLU, Dropout, BatchNormalization, SpatialDropout1D, concatenate, Concatenate
from keras.layers import Embedding, Dense, Dropout, Bidirectional, CuDNNLSTM
from keras.layers import Add, Subtract, Multiply, Dot
from attention import Attention
from keras import initializers, regularizers, constraints, optimizers, layers
from keras.preprocessing.sequence import pad_sequences
from keras.utils import Sequence
from keras.callbacks import Callback
from keras import backend as K
from keras.engine.topology import Layer 
# from keras.utils.training_utils import multi_gpu_model

tic = time.time()
DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'
WEIGHT_PATH = './weights/'
if not os.path.exists(SAVE_PATH):
    print('create dir: %s' % SAVE_PATH)
    os.mkdir(SAVE_PATH)
if not os.path.exists(WEIGHT_PATH):
    print('create dir: %s' % WEIGHT_PATH)
    os.mkdir(WEIGHT_PATH)

print('ques_user_title_desc_textattbirnn_cossim_enc...')

data = pd.read_pickle(os.path.join(DATA_PATH, 'invite_data.pkl'))
prev_ans_ques_title_desc = pd.read_pickle(os.path.join(DATA_PATH, 'prev_ans_ques_title_desc.pkl'))
word = pd.read_pickle(os.path.join(DATA_PATH, 'word.pkl'))
len_train = (~data['label'].isnull()).sum()
print(len_train)

data.head()

prev_ans_ques_title_desc.head()

data = pd.concat([data, prev_ans_ques_title_desc], axis=1)
data.head()

max_len_tq = data['title_q'].apply(len).max()
max_len_dq = data['desc_q'].apply(len).max()
max_len_tu = data['title_u'].apply(len).max()
max_len_du = data['desc_u'].apply(len).max()
print(max_len_tq, max_len_dq) # 110
print(max_len_tu, max_len_du) # 110

max_len_tq = 38
max_len_dq = 256
max_len_tu = 64
max_len_du = 256

np.random.seed(42)
def shuffle(l):
    l = l[:256]
    np.random.shuffle(l)
    return l

data['title_q'] = data['title_q'].progress_apply(shuffle)
data['desc_q'] = data['desc_q'].progress_apply(shuffle)
data['title_u'] = data['title_u'].progress_apply(shuffle)
data['desc_u'] = data['desc_u'].progress_apply(shuffle)

data.head()

title_q = pad_sequences(data['title_q'], maxlen=max_len_tq)
desc_q = pad_sequences(data['desc_q'], maxlen=max_len_dq)
title_u = pad_sequences(data['title_u'], maxlen=max_len_tu)
desc_u = pad_sequences(data['desc_u'], maxlen=max_len_du)

title_q.shape

desc_q.shape

# embed_weights = np.array([[0] * 64] + [list(v) for v in word['vector'].values])
embed_weights = np.vstack([np.zeros((1, 64))] + list(word['vector'].values))

embed_weights.shape

del prev_ans_ques_title_desc
gc.collect()

class DataSequence(Sequence):
    
    def __init__(self, tq, tu, dq, du, y, batch_size=128):
        self.tq = tq
        self.tu = tu
        self.dq = dq
        self.du = du
        self.y = y
        self.batch_size = batch_size
         
    def __len__(self):
        return int(np.ceil(self.tq.shape[0] / float(self.batch_size)))
    
    def __getitem__(self, idx):
        batch_idx = np.arange(idx * self.batch_size, min((idx + 1) * self.batch_size, self.tq.shape[0]))
        batch_tq = self.tq[batch_idx]
        batch_tu = self.tu[batch_idx]
        batch_dq = self.dq[batch_idx]
        batch_du = self.du[batch_idx]
        batch_y = self.y[batch_idx]
        return [batch_tq, batch_tu, batch_dq, batch_du], batch_y

class MetricsCallback(Callback):
    def __init__(self, trn_tq, trn_tu, trn_dq, trn_du, y_trn,
                       val_tq, val_tu, val_dq, val_du, y_val, 
                       batch_size=128, save_name='weight.h5'):
        self.trn_generator = DataSequence(trn_tq, trn_tu, trn_dq, trn_du, y_trn, BATCH_SIZE)
        self.val_generator = DataSequence(val_tq, val_tu, val_dq, val_du, y_val, BATCH_SIZE)
        self.y_trn = y_trn
        self.y_val = y_val
        self.save_name = save_name
        self.best_score = 0.5

    def on_epoch_end(self, epoch, logs={}):
        # eval train
        y_pred = self.model.predict_generator(self.trn_generator, 
                                              max_queue_size=10, 
                                              workers=1, 
                                              use_multiprocessing=False, 
                                              verbose=0)
        roc = roc_auc_score(self.y_trn, y_pred)
        # eval valid
        y_pred_val = self.model.predict_generator(self.val_generator, 
                                              max_queue_size=10, 
                                              workers=1, 
                                              use_multiprocessing=False, 
                                              verbose=0)
        roc_val = roc_auc_score(self.y_val, y_pred_val)
        print('\rroc-auc: %s - roc-auc_val: %s' % (str(round(roc,4)),str(round(roc_val,4))),end=100*' '+'\n')
#         print(roc_val)
#         print(log_loss(self.y_val, y_pred_val)/len(y_pred_val))
        if roc_val > self.best_score:
            self.best_score = roc_val
            self.model.save_weights(os.path.join(WEIGHT_PATH, self.save_name))
        
        return

def Net():
    # title q
    inp_tq = Input(shape=(max_len_tq,))
    x1 = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_tq)
    x1 = Bidirectional(CuDNNLSTM(128, return_sequences=True))(x1)  # LSTM or GRU
    x1 = Attention(max_len_tq)(x1)
    x1 = Dense(256)(x1)
    x1 = BatchNormalization()(x1)
    x1 = PReLU()(x1)
    x1 = Dropout(0.2)(x1)
    # title u
    inp_tu = Input(shape=(max_len_tu,))
    x2 = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_tu)
    x2 = Bidirectional(CuDNNLSTM(128, return_sequences=True))(x2)  # LSTM or GRU
    x2 = Attention(max_len_tu)(x2)
    x2 = Dense(256)(x2)
    x2 = BatchNormalization()(x2)
    x2 = PReLU()(x2)
    x2 = Dropout(0.2)(x2)
    # desc q
    inp_dq = Input(shape=(max_len_dq,))
    y1 = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_dq)
    y1 = Bidirectional(CuDNNLSTM(128, return_sequences=True))(y1)  # LSTM or GRU
    y1 = Attention(max_len_dq)(y1)
    y1 = Dense(256)(y1)
    y1 = BatchNormalization()(y1)
    y1 = PReLU()(y1)
    y1 = Dropout(0.2)(y1)
    # desc u
    inp_du = Input(shape=(max_len_du,))
    y2 = Embedding(embed_weights.shape[0], embed_weights.shape[1], weights=[embed_weights], trainable=False)(inp_du)
    y2 = Bidirectional(CuDNNLSTM(128, return_sequences=True))(y2)  # LSTM or GRU
    y2 = Attention(max_len_du)(y2)
    y2 = Dense(256)(y2)
    y2 = BatchNormalization()(y2)
    y2 = PReLU()(y2)
    y2 = Dropout(0.2)(y2)
    ## cossim
    cs1 = Dot(axes=-1,normalize=True)([x1, x2])
    cs2 = Dot(axes=-1,normalize=True)([y1, y2])
    cs = Concatenate(axis=-1)([cs1, cs2])
    out = Dense(1, activation="sigmoid")(cs)
    model = Model(inputs=[inp_tq, inp_tu, inp_dq, inp_du], outputs=out)
    return model

train_tq, train_tu, train_dq, train_du = title_q[:len_train], title_u[:len_train], desc_q[:len_train], desc_u[:len_train]
test_tq, test_tu, test_dq, test_du = title_q[len_train:], title_u[len_train:], desc_q[len_train:], desc_u[len_train:]
train_y = data['label'].values[:len_train]

print(train_tq.shape)
print(test_tq.shape)
print(train_y.sum() / len(train_y))

# BATCH_SIZE = 1024

# kfold = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

# for i, (tr_idx, va_idx) in enumerate(kfold.split(train_tq, train_y)):
    
#     print('-'*100)
#     print('Fold %d' % i)
#     trn_tq, trn_tu, trn_dq, trn_du, trn_y = train_tq[tr_idx], train_tu[tr_idx], train_dq[tr_idx], train_du[tr_idx], train_y[tr_idx]
#     val_tq, val_tu, val_dq, val_du, val_y = train_tq[va_idx], train_tu[va_idx], train_dq[va_idx], train_du[va_idx], train_y[va_idx]
    
#     K.clear_session()
#     model = Net()
#     model.compile(loss='binary_crossentropy', optimizer='adam',)
    
#     trn_generator = DataSequence(trn_tq, trn_tu, trn_dq, trn_du, trn_y, batch_size=BATCH_SIZE)
#     val_generator = DataSequence(val_tq, val_tu, val_dq, val_du, val_y, batch_size=BATCH_SIZE)
    
#     history = model.fit_generator(generator=trn_generator, 
#                         epochs=10, 
#                         verbose=1, 
#                         callbacks=[MetricsCallback(trn_tq, trn_tu, trn_dq, trn_du, trn_y,
#                                                    val_tq, val_tu, val_dq, val_du, val_y, 
#                                                    batch_size=BATCH_SIZE, 
#                                                    save_name='ques_user_title_desc_textattbirnn_cossim_enc_weight_fold%d.h5' % i)], 
#                         max_queue_size=100, 
#                         workers=1, 
#                         use_multiprocessing=False,)
    
#     del trn_tq, trn_tu, trn_dq, trn_du, trn_y
#     del val_tq, val_tu, val_dq, val_du, val_y
#     gc.collect()

# del trn_tq, trn_tu, trn_dq, trn_du, trn_y
# del val_tq, val_tu, val_dq, val_du, val_y
# gc.collect()

# 0.4608

BATCH_SIZE = 256

kfold = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

test_generator = DataSequence(test_tq, test_tu, test_dq, test_du, np.zeros(test_tq.shape[0]), batch_size=BATCH_SIZE)

test_pred = np.zeros((test_tq.shape[0], 1))
train_pred = np.zeros((train_tq.shape[0], 1))

for i, (tr_idx, va_idx) in enumerate(kfold.split(train_tq, train_y)):

    print('-'*100)
    print('Fold %d' % i)
    val_tq, val_tu, val_dq, val_du, val_y = train_tq[va_idx], train_tu[va_idx], train_dq[va_idx], train_du[va_idx], train_y[va_idx]

    K.clear_session()
    model = Net()
    model.compile(loss='binary_crossentropy', optimizer='adam',)

    val_generator = DataSequence(val_tq, val_tu, val_dq, val_du, val_y, batch_size=BATCH_SIZE)

    model.load_weights(os.path.join(WEIGHT_PATH, 'ques_user_title_desc_textattbirnn_cossim_enc_weight_fold%d.h5' % i))

    train_pred[va_idx] = model.predict_generator(val_generator, verbose=1)
    test_pred += model.predict_generator(test_generator, verbose=1) / 5
    
    print(roc_auc_score(val_y, train_pred[va_idx]))
    
    del val_tq, val_tu, val_dq, val_du, val_y
    gc.collect()
    print('Predict Done.')

# del trn_tq, trn_tu, trn_dq, trn_du, trn_y
# del val_tq, val_tu, val_dq, val_du, val_y
# gc.collect()

arr = np.vstack([train_pred, test_pred])
feat = pd.DataFrame(arr, columns=['ques_user_title_desc_textattbirnn_cossim_enc'])
feat.head()

feat.to_pickle('./feats/ques_user_title_desc_textattbirnn_cossim_enc.pkl')

feat.shape

!pip list | grep scikit
