==> 40.gen_cbt_data.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import gc
import time
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold
from catboost import CatBoostClassifier, Pool
import pickle
import os

DATA_PATH = '../pkl/'
SAVE_PATH = './feats/'

print('load data...')
data = pd.read_pickle(os.path.join(DATA_PATH, 'invite_data.pkl'))

feats = ['inv_ans_gap', 'user_feat', 'question_feat', 'user_question_feat',
         'global_user_ques_expo', 'local_user_ques_expo', 'prev_ans_feat', 'prev_ans_times_st',
         'inv_times_st', 'history_user_ques_accept', 'topic_qu_sims_stat', 'prev_topic_sims_stat',
         'ques_topic_dnn_enc_A', 'ques_topic_dnn_enc_B', 'ques_topic_dnn_enc_C',
         'user_topic_dnn_enc_A', 'user_topic_dnn_enc_B', 'user_topic_dnn_enc_C',
         'user_ques_topic_dnn_enc_A', 'user_ques_topic_dnn_enc_B', 'user_ques_topic_dnn_enc_C',
         'title_desc_textcnn_enc', 'title_desc_textattbirnn_enc',
         'ques_user_title_desc_textattbirnn_enc', 
         'ques_user_title_desc_textattbirnn_cossim_enc',
         'user_ques_title_desc_com_stat', 'user_ques_title_desc_sims_stat'
        ]

# 'local_ques_expo_change', 'local_user_expo_change', 'ques_user_title_desc_textattbirnn_enc_no_shuffle', 'ques_user_title_desc_textcnn_enc'

feats_dat = []
for f in feats:
    feats_dat.append(pd.read_pickle(os.path.join(SAVE_PATH, '%s.pkl' % f)))

print('concat feat...')
print(data.shape)
data = pd.concat([data] + feats_dat, axis=1)
print(data.shape)

print(list(data.columns))

data.to_pickle(os.path.join(DATA_PATH, 'cbt_data.pkl'))
